Update db.js
Ton of changes. Old file should be dismissed. This changes the object functions to prototype variables and most of the methods I've tested work now (Or at least they don't throw errors anymore). Be sure that all boolean values are entered as 0 or 1 (0 being false). The main issue that remains is in closing the database -- When I close the DB on the first form and then use require on the modify form, I get some exception about re-opening an already closed database, which evidently you can't do? I've removed my close() call in myRecipes and it seems to work.