Create ModifyFormHelper.js
Here is the Entity logic and how it can be implemented. I'm not certain about the wiring in that form so you'll have to stick these where they should go, but that should at least show you some information about the scaling. If it isn't working, upload it as a new file and email me and I can try to debug to see what is wrong with the logic.